refactor(sicklegraph-plugin): reorganize service imports and move SickleGraphService to a new file
