+++ src/sicklegraph-plugin/services/sickle-graph-service.ts
